requirement.txt
absl-py==1.4.0
aiohttp==3.8.5
aiosignal==1.3.1
annotated-types==0.5.0
astunparse==1.6.3
async-timeout==4.0.2
attrs==23.1.0
Automat==22.10.0
backcall==0.2.0
beautifulsoup4==4.12.2
blinker==1.6.2
cachetools==5.3.1
certifi==2023.5.7
cffi==1.15.1
chardet==3.0.4
charset-normalizer==3.1.0
click==8.1.3
colorama==0.4.6
comtypes==1.1.14
constantly==15.1.0
contourpy==1.1.0
cryptography==41.0.3
cssselect==1.2.0
cycler==0.11.0
decorator==4.4.2
distlib==0.3.6
EasyProcess==1.1
entrypoint2==1.1
et-xmlfile==1.1.0
executing==1.2.0
filelock==3.12.2
Flask==2.3.2
flatbuffers==23.5.26
fonttools==4.41.1
frozenlist==1.4.0
gast==0.4.0
google-auth==2.22.0
google-auth-oauthlib==1.0.0
google-pasta==0.2.0
googletrans==3.0.0
grpcio==1.57.0
gTTS==2.3.2
h11==0.9.0
h2==3.2.0
h5py==3.9.0
hpack==3.0.0
hstspreload==2023.1.1
htmlmin==0.1.12
httpcore==0.9.1
httpx==0.13.3
hyperframe==5.2.0
hyperlink==21.0.0
idna==2.10
ImageHash==4.3.1
imageio==2.31.1
imageio-ffmpeg==0.4.8
importlib-metadata==6.7.0
incremental==22.10.0
itemadapter==0.8.0
itemloaders==1.1.0
itsdangerous==2.1.2
Jinja2==3.1.2
jmespath==1.0.1
joblib==1.1.1
keras==2.13.1
kiwisolver==1.4.4
libclang==16.0.6
lxml==4.9.3
Markdown==3.4.4
MarkupSafe==2.1.2
matplotlib==3.7.2
missingno==0.5.2
MouseInfo==0.1.3
moviepy==1.0.3
mss==9.0.1
multidict==6.0.4
multimethod==1.9.1
mysql-connector-python==8.0.33
networkx==3.1
numpy==1.24.3
oauthlib==3.2.2
openai==0.27.8
opencv-python==4.7.0.72
openpyxl==3.1.2
opt-einsum==3.3.0
packaging==23.1
pandas==2.0.3
pandas-profiling==3.2.0
parsel==1.8.1
phik==0.12.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.5.3
playsound==1.2.2
proglog==0.1.10
Protego==0.2.1
protobuf==3.20.3
pure-eval==0.2.2
pyasn1==0.5.0
pyasn1-modules==0.3.0
PyAutoGUI==0.9.54
pycparser==2.21
pydantic==1.8.2
pydantic_core==2.4.0
PyDispatcher==2.0.7
PyGetWindow==0.0.9
Pygments==2.16.1
PyMsgBox==1.0.9
pyOpenSSL==23.2.0
pyparsing==3.0.9
pyperclip==1.8.2
pypiwin32==223
PyRect==0.2.0
pyscreenshot==3.1
PyScreeze==0.1.29
python-dateutil==2.8.2
pyttsx3==2.90
pytube==15.0.0
pytweening==1.0.7
pytz==2023.3
PyWavelets==1.4.1
pywhatkit==5.4
pywin32==306
PyYAML==6.0.1
pyzmq==25.1.1
queuelib==1.6.2
requests==2.31.0
requests-file==1.5.1
requests-oauthlib==1.3.1
rfc3986==1.5.0
rsa==4.9
scikit-learn==1.3.0
scipy==1.11.1
Scrapy==2.10.0
seaborn==0.12.2
service-identity==23.1.0
six==1.16.0
sklearn==0.0.post7
sniffio==1.3.0
soupsieve==2.4.1
tangled-up-in-unicode==0.2.0
tensorboard==2.13.0
tensorboard-data-server==0.7.1
tensorflow==2.13.0
tensorflow-estimator==2.13.0
tensorflow-intel==2.13.0
tensorflow-io-gcs-filesystem==0.31.0
termcolor==2.3.0
threadpoolctl==3.2.0
tldextract==3.4.4
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
Twisted==22.10.0
twisted-iocpsupport==1.0.3
typing_extensions==4.5.0
tzdata==2023.3
urllib3==1.26.16
vboxapi==1.0
ve==0.0.1
virtualenv==20.23.0
visions==0.7.4
w3lib==2.1.2
wcwidth==0.2.6
Werkzeug==2.3.4
wikipedia==1.4.0
wrapt==1.15.0
yarl==1.9.2
youtube-dl==2021.12.17
zipp==3.15.0
zope.interface==6.0
